modify QEventDispatcherGlib::processEvents to match documentation
processEvents(flags) without a timeout is documented to process
all pending events, but not events which are created during
this processing. For the glib backend, this was not the case.
In the presence of high priority events, only high priority events
were processed. This patch adapts the implementation to the
documentation.

Fixes: QTBUG-84291
Fixes: QTBUG-87137
Fixes: QTBUG-87307
Change-Id: I0bc3c6a5cd671c4d1e2a90f61637b62bf3a99c2c